server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-pptx
pydantic
requests
openai
